Refactor: re-export record formatters from index
