+++ src/handlers/tool-configs/index.ts
@@ -5,7 +5,7 @@ export { companyToolConfigs, companyToolDefinitions } from './companies/index.js
 export { peopleToolConfigs, peopleToolDefinitions } from './people.js';
 export { listsToolConfigs, listsToolDefinitions } from './lists.js';
 export { promptsToolConfigs, promptsToolDefinitions } from './prompts.js';
-export { recordToolConfigs, recordToolDefinitions } from './records.js';
+export { recordToolConfigs, recordToolDefinitions } from './records/index.js';
 export { paginatedPeopleToolConfigs, paginatedPeopleToolDefinitions } from './paginated-people.js';
 export { tasksToolConfigs, tasksToolDefinitions } from './tasks.js';
 export {
